rudof_rdf::rdf_core

Module term

Modules

literal

Structs

ConcreteBlankNode

A lightweight borrowed representation of a blank node.

ConcreteTriple

A concrete implementation of an RDF triple for a specific RDF model.

Enums

IriOrBlankNode

Represents an RDF resource that is either an IRI or a blank node.

Object

Represents an RDF object value in the object position of a triple.

TermKind

Represents the kind of RDF term.

Traits

BlankNode

Represents an owned blank node in RDF.

BlankNodeRef

A trait for borrowed references to blank node labels.

Iri

Represents an IRI (Internationalized Resource Identifier) in RDF.

Subject

Represents the subject position of an RDF triple.

Term

Represents any RDF term that can appear in an RDF graph.

Triple

Represents an RDF triple.