rudof_rdf::rdf_core::visualizer

Module style

Structs

ArrowStyle

Defines the visual style of an arrow in a PlantUML diagram.

StereotypeStyle

Represents the visual style of a UML stereotype in PlantUML.

Style

Main style container for RDF visualization.

Enums

ThicknessStyle

Represents the available line thickness and style options in PlantUML.

UmlColor

Represents the set of colors supported by PlantUML.

Constants

DEFAULT_OBJECT_ARROW_STYLE

Default style for object arrows.

DEFAULT_PREDICATE_ARROW_STYLE

Default style for predicate arrows.

DEFAULT_SUBJECT_ARROW_STYLE

Default style for subject arrows.