rudof_rdf

Module rdf_impl

Structs

InMemoryGraph

An RDF graph stored entirely in memory.

SparqlEndpoint

A SPARQL endpoint client that implements the SRDF interface.

SparqlVars

Utility struct for displaying SPARQL variable lists.

Enums

InMemoryGraphError

Represents all possible errors that can occur when working with in-memory RDF graphs.

ReaderMode

Reader mode when parsing RDF data files.

SparqlEndpointError

Represents all possible errors that can occur when interacting with SPARQL endpoints.